+++ CHANGES.md
@@ -54,7 +54,8 @@ Beta Releases
 * Added prototype `clone` and `merge` functions to `DynamicScene` graphics objects .
 * Added `width`, `height`, and `nearFarScalar` properties to `DynamicBillboard` for controlling the image size.
 * Added support for complex polygons. This fixes [#1121](https://github.com/AnalyticalGraphicsInc/cesium/issues/1121).
-  
+* Added a new `GeocodingWidget` that allows users to enter an address or the name of a landmark and zoom the camera to that location.  It is enabled by default in applications that use the `Viewer`.
+
 ### b20 - 2013-09-03
 
 _This releases fixes 2D and other issues with Chrome 29.0.1547.57 ([#1002](https://github.com/AnalyticalGraphicsInc/cesium/issues/1002) and [#1047](https://github.com/AnalyticalGraphicsInc/cesium/issues/1047))._